abhi-ai-portfolio/projects/llm-qa-bot/requirements.txt
gradio
transformers
torch
sentence-transformers
faiss-cpu
